backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# NLP and ML
sentence-transformers==2.2.2
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Vector Database
qdrant-client==1.7.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0
